// backend/src/types/api.types.ts
export interface AuthRequest extends FastifyRequest {
  user?: {
    id: string
    email: string
    aud: string
    role?: string
  }
}